[Codegen] Remove dependencies on Flow
