+++ compiler/src/iree/compiler/Codegen/Common/BufferizationAnalysis.cpp
@@ -14,8 +14,6 @@
 #include "iree/compiler/Codegen/Common/BufferizationAnalysis.h"
 
 #include "iree/compiler/Codegen/Utils/Utils.h"
-#include "iree/compiler/Dialect/Flow/IR/FlowOps.h"
-#include "iree/compiler/Dialect/Flow/IR/FlowTypes.h"
 #include "iree/compiler/Dialect/HAL/IR/HALOps.h"
 #include "iree/compiler/Dialect/LinalgExt/IR/LinalgExtOps.h"
 #include "iree/compiler/Dialect/TensorExt/IR/TensorExtOps.h"
