+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/BUILD.bazel
@@ -71,7 +71,6 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Codegen/Interfaces:BufferizationInterfaces",
         "//compiler/src/iree/compiler/Codegen/Transforms",
         "//compiler/src/iree/compiler/Codegen/Utils",
-        "//compiler/src/iree/compiler/Dialect/Flow/IR",
         "//compiler/src/iree/compiler/Dialect/HAL/IR",
         "//compiler/src/iree/compiler/Dialect/LinalgExt/IR",
         "//llvm-external-projects/iree-dialects:IREEDialectsTransforms",
