+++ compiler/src/iree/compiler/Codegen/Common/TransformExtensions/CMakeLists.txt
@@ -74,7 +74,6 @@ iree_cc_library(
     iree::compiler::Codegen::Interfaces::BufferizationInterfaces
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
-    iree::compiler::Dialect::Flow::IR
     iree::compiler::Dialect::HAL::IR
     iree::compiler::Dialect::LinalgExt::IR
   PUBLIC
