+++ compiler/src/iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -190,8 +190,6 @@ iree_cc_library(
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
     iree::compiler::Codegen::Utils::VectorOpUtils
-    iree::compiler::Dialect::Flow::IR
-    iree::compiler::Dialect::Flow::Transforms
     iree::compiler::Dialect::HAL::IR
     iree::compiler::Dialect::HAL::Transforms
     iree::compiler::Dialect::LinalgExt::IR
