+++ compiler/src/iree/compiler/Codegen/WGSL/BUILD.bazel
@@ -54,7 +54,6 @@ iree_compiler_cc_library(
     deps = [
         ":PassHeaders",
         ":PassesIncGen",
-        "//compiler/src/iree/compiler/Dialect/Flow/IR",
         "//compiler/src/iree/compiler/Dialect/HAL/IR",
         "//compiler/src/iree/compiler/Dialect/HAL/IR:HALDialect",
         "@llvm-project//mlir:ArithDialect",
