+++ compiler/src/iree/compiler/Dialect/TensorExt/Transforms/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/Transforms/BUILD.bazel                    #
+# compiler/src/iree/compiler/Dialect/TensorExt/Transforms/BUILD.bazel          #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -16,33 +16,12 @@ iree_cc_library(
   HDRS
     "Transforms.h"
   SRCS
-    "AffineMinDistributedSCFCanonicalization.cpp"
-    "RemoveSingleIterationLoop.cpp"
-    "Transforms.cpp"
+    "Folders.cpp"
   DEPS
     LLVMSupport
     MLIRAffineDialect
-    MLIRAffineTransforms
     MLIRAffineUtils
-    MLIRAnalysis
-    MLIRArithDialect
-    MLIRFuncDialect
-    MLIRFunctionInterfaces
-    MLIRIR
-    MLIRLinalgDialect
-    MLIRLinalgTransforms
-    MLIRMemRefDialect
-    MLIRPass
-    MLIRSCFTransforms
-    MLIRSupport
     MLIRTensorDialect
-    MLIRTransformUtils
-    MLIRTransforms
-    MLIRValueBoundsOpInterface
-    MLIRVectorDialect
-    iree::compiler::Codegen::Dialect::Codegen::IR::IREECodegenDialect
-    iree::compiler::Codegen::Utils
-    iree::compiler::Dialect::HAL::IR
     iree::compiler::Dialect::TensorExt::IR
   PUBLIC
 )
